Starting work
I have finished adding the text and lines in the request, and updated some older text to make the example better, and added in logic for multiple line description. Splitting the logic away from the GUI for easier editing and to prepare for alternative inputs. There was a possible issue with all the rendering code being repeated for each page, I put the code needed for each page in another loop for now but this issue remains unaddressed.